Applied patch from Peter Green to build with -O2 on ia64
From: peter green
To: Bug 504087

I have attatched a patch which reduces the optimisation level of the
build to O2 on ia64.   I have checked the patch sets the optimisation
level correctly by using a different architecture in the test but I do
not have access to an ia64 box so I can't test to confirm the
subitters
statement that doing so fix the build on ia64.